Revert "Feature/player attack"
