--- HorseMadh/Assets/Prefabs/AI-Horsey.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: fa7ed0f52fca26741b9ce6459a82e5a4
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
