--- HorseMadh/Assets/Scripts/Inputs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f6462802dbaf68d4a802f4c36477a710
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
