--- HorseMadh/Assets/Scripts/Inputs/AiInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e5c10abf06b4d5b4abaa377b3a96d757
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
