--- HorseMadh/Assets/Scripts/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 62abb511cb08e304892f5a3fa4eff0f1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
