--- HorseMadh/Assets/Scripts/Player/PlayerAttack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 013abf0dbbe88f048bbdda72ad37b9fb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
