--- HorseMadh/Assets/Scripts/SpriteLookAtCam.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ce3558471ece8f54092b0e205a2cbae8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
